requirements.txt
langchain 
langchain_community
streamlit 
dotenv
pymupdf
sentence-transformers
faiss-cpu
langchain_huggingface